+++ example/transactions/applepay_redirect.php
@@ -8,20 +8,13 @@ $buckaroo = new BuckarooClient($_ENV['BPE_WEBSITE_KEY'], $_ENV['BPE_SECRET_KEY']
 
 //Also accepts json
 //Pay
-$response = $buckaroo->method('applepay')->pay([
-    'amountDebit' => 10,
-    'invoice' => uniqid(),
-    'paymentData' => uniqid(),
-    'customerCardName'  => 'Buck Aroo'
-]);
-
-//Pay Redirect
 $response = $buckaroo->method('applepay')->payRedirect([
     'amountDebit' => 10,
     'invoice' => uniqid(),
     'servicesSelectableByClient' => 'applepay',
-    'continueOnIncomplete' => '1',
+    'continueOnIncomplete' => '1'
 ]);
+
 
 //Refund
 $response = $buckaroo->method('applepay')->refund([