+++ example/transactions/giropay.php
@@ -9,6 +9,7 @@ $buckaroo = new BuckarooClient($_ENV['BPE_WEBSITE_KEY'], $_ENV['BPE_SECRET_KEY']
 //Also accepts json
 //Pay
 $response = $buckaroo->method('giropay')->pay([
+    'invoice' => uniqid(),
     'bic' => 'GENODETT488',
     'amountDebit' => 10.10
 ]);