+++ example/transactions/kbc.php
@@ -9,6 +9,7 @@ $buckaroo = new BuckarooClient($_ENV['BPE_WEBSITE_KEY'], $_ENV['BPE_SECRET_KEY']
 //Also accepts json
 //Pay
 $response = $buckaroo->method('kbcpaymentbutton')->pay([
+    'invoice' => uniqid(),
     'amountDebit' => 10.10
 ]);