+++ example/transactions/tinka.php
@@ -3,60 +3,72 @@
 require('../bootstrap.php');
 
 use Buckaroo\BuckarooClient;
+use Buckaroo\Resources\Constants\Gender;
 
 $buckaroo = new BuckarooClient($_ENV['BPE_WEBSITE_KEY'], $_ENV['BPE_SECRET_KEY']);
 
 //Also accepts json
 //Pay
 $response = $buckaroo->method('tinka')->pay([
-    'amountDebit'       => 3.5,
-    'order'             => uniqid(),
-    'invoice'           => uniqid(),
-    'description'       => 'This is a test order',
-    'serviceParameters' => [
-        'paymentMethod'       => 'Credit',
-        'deliveryMethod'      => 'Locker',
-        'deliveryDate'          => '09-12-2022',
-        'articles'      => [
-            [
-                'type'              => 1,
-                'description'       => 'Blue Toy Car',
-                'brand'             => 'Ford Focus',
-                'manufacturer'      => 'Ford',
-                'color'             => 'Red',
-                'size'              => 'Small',
-                'quantity'          => '1',
-                'grossUnitPrice'    => '3.5',
-                'unitCode'         => 'test'
-            ]
+    'amountDebit'           => 3.5,
+    'order'                 => uniqid(),
+    'invoice'               => uniqid(),
+    'description'           => 'This is a test order',
+    'paymentMethod'         => 'Credit',
+    'deliveryMethod'        => 'Locker',
+    'deliveryDate'          => '2030-01-01',
+    'articles'              => [
+        [
+            'type'              => 1,
+            'description'       => 'Blue Toy Car',
+            'brand'             => 'Ford Focus',
+            'manufacturer'      => 'Ford',
+            'color'             => 'Red',
+            'size'              => 'Small',
+            'quantity'          => '1',
+            'price'             => '3.5',
+            'unitCode'         => 'test'
+        ]
+    ],
+    'customer'          => [
+        'gender'        => Gender::MALE,
+        'firstName'     => 'Buck',
+        'lastName'      => 'Aroo',
+        'initials'       => 'BA',
+        'birthDate'     => '1990-01-01',
+    ],
+    'billing'              => [
+        'recipient'          => [
+            'lastNamePrefix'    => 'the'
         ],
-        'customer'      => [
-            'gender'        => '1',
-            'initials'      => 'J.S.',
-            'firstName' => 'Test',
-            'lastName' => 'Aflever',
-            'birthDate' => '01-01-1990',
-            'billing'                   => [
-                'prefixLastName'    => 'the',
-                'email' => 'billingcustomer@buckaroo.nl',
-                'phone' => '0109876543',
-                'street' => 'Hoofdstraat',
-                'streetNumber' => '80',
-                'streetNumberAdditional' => 'A',
-                'postalCode' => '8441EE',
-                'city' => 'Heerenveen',
-                'country' => 'NL',
-            ],
-            'shipping'                  => [
-                'externalName' => 'Test',
-                'phone' => '0109876543',
-                'street' => 'Hoofdstraat',
-                'streetNumber' => '80',
-                'streetNumberAdditional' => 'A',
-                'postalCode' => '8441EE',
-                'city' => 'Heerenveen',
-                'country' => 'NL',
-            ]
+        'email' => 'billingcustomer@buckaroo.nl',
+        'phone' => [
+            'mobile' => '0109876543',
+        ],
+        'address'                   => [
+            'street' => 'Hoofdstraat',
+            'houseNumber' => '80',
+            'houseNumberAdditional' => 'A',
+            'zipcode' => '8441EE',
+            'city' => 'Heerenveen',
+            'country' => 'NL',
+        ]
+    ],
+    'shipping'              => [
+        'recipient'          => [
+            'lastNamePrefix'    => 'the'
+        ],
+        'email' => 'billingcustomer@buckaroo.nl',
+        'phone' => [
+            'mobile' => '0109876543',
+        ],
+        'address'                   => [
+            'street' => 'Hoofdstraat',
+            'houseNumber' => '80',
+            'houseNumberAdditional' => 'A',
+            'zipcode' => '8441EE',
+            'city' => 'Heerenveen',
+            'country' => 'NL',
         ]
     ]
 ]);