+++ src/PaymentMethods/BuckarooVoucher/Models/Create.php
@@ -18,18 +18,17 @@
  * @license   https://tldrlegal.com/license/mit-license
  */
 
-namespace Buckaroo\PaymentMethods\GiftCard\Models;
+namespace Buckaroo\PaymentMethods\BuckarooVoucher\Models;
 
 use Buckaroo\Models\ServiceParameter;
 
-class Pay extends ServiceParameter
+class Create extends ServiceParameter
 {
-    protected string $fashionChequeCardNumber;
-    protected string $fashionChequePin;
-    protected string $intersolveCardnumber;
-    protected string $intersolvePIN;
-    protected string $tcsCardnumber;
-    protected string $tcsValidationCode;
-    protected string $lastName;
-    protected string $email;
+    protected string $usageType;
+
+    protected string $validFrom;
+
+    protected string $validUntil;
+
+    protected string $creationBalance;
 }