+++ src/PaymentMethods/BuckarooVoucher/Models/Pay.php
@@ -18,18 +18,11 @@
  * @license   https://tldrlegal.com/license/mit-license
  */
 
-namespace Buckaroo\PaymentMethods\GiftCard\Models;
+namespace Buckaroo\PaymentMethods\BuckarooVoucher\Models;
 
 use Buckaroo\Models\ServiceParameter;
 
 class Pay extends ServiceParameter
 {
-    protected string $fashionChequeCardNumber;
-    protected string $fashionChequePin;
-    protected string $intersolveCardnumber;
-    protected string $intersolvePIN;
-    protected string $tcsCardnumber;
-    protected string $tcsValidationCode;
-    protected string $lastName;
-    protected string $email;
+    protected string $vouchercode;
 }